Add sine wave generator
Sine wave generator is based on an infinite impulse response
(IIR) filter. Generating the next sample requires a single
floating-point multiply-add, which is extremely efficient
compared to the usual method of calling system sin().

A sin() function is still required to set it up, so we also
include the same one in BSD libc (more or less), allowing us
to drop the dependency on <math.h> while maintaing accuracy.