Fix builds against ancient versions of PipeWire
struct pw_buffer::requested only exists in PipeWire 0.3.50
or later. Add a workaround for not having it that should
work on older versions at the cost(?) of some latency.

Part of this workaround is ensuring that we never try to
generate more samples than the sample generator callback
output buffer can hold. This applies to PulseAudio as well.

Another holdup is (was) PW_KEY_NODE_RATE, added in PipeWire
0.3.33 to support dynamic node rates. We don't really need to
specify it for our simple use case, so we may stop doing so.