Fix memory leaks in PipeWire and ALSA found with ASan
For PipeWire, we forgot to call pw_deinit(). That's fair.

For ALSA, we didn't call snd_config_update_free_global().
It's not obvious that snd_pcm_open() touches a global that
snd_pcm_close() doesn't, but it's fixed on real ALSA now.
(There are still leaks on PipeWire's ALSA emulation.)

Unfortunately, on PipeWire's PulseAudio emulation,
a segfault occurs if PipeWire itself is not running.
Let's hope this doesn't happen on real PulseAudio.